Ensure eigenvalues are positive in SPD matrices (#372)
By definition symmetric positive definite matrices have positive
eigenvalues. Numerical errors may result in infinitesimal negative
eigenvalues being computed, resulting in DomainErrors when the sqrt or
log of the eigenvalue is calculated.

This change replaces any negative eigenvalues with the smallest positive
value to ensure that the sqrt is successful.

This makes it the user's responsibility to ensure that input arguments
are valid SPDs before passing.