special_euclidean: Remove specific log/exp implementations
These implementations are the one from *group product*,
so they are not invariant with respect to the semidirect
product.

One could put them back in the product_group layer instead.
The proper way to invoke them is then `exp(base_manifold(G), ...)`
instead of `exp(G, ...)`.

Until this is fixed, `exp` now uses more allocations, even
when calling it with `base_manifold`.