Removed expicit complex conjugate in Stiefel.jl documentation
Removed, because it should not be there. It now matches the description given in Generalized Stiefel. See below for a short proof. 

For some interval, $I=[a,b]$ where $a,b\in \mathbb{R}$, the parametrized curve is the smooth map $c\colon I\to \mathcal{M}$. It is $c(t)\in \operatorname{St}(n,k)$ where $c(0)=p$ and, the velocity of the curve is $c'(0)=X$. Since $c(t)$ is in $\operatorname{St}(n,k)$, it must satisfy
$c(t)^{\mathrm{H}}c(t)=I_{k}$ 
for all $t\in I$. Taking the derivative of with respect to $t$ at $t=0$ the expression becomes
$c'(t)^{\mathrm{H}}c(t)+c(t)^{\mathrm{H}}c'(t)=X^{\mathrm{H}}p+p^{\mathrm{H}}X=0_{k}.$
This means that the tangent space of $\operatorname{St}(n,k)$ at the point $p$ is
$\operatorname{T}_{p}\operatorname{St}(n,k)=\left\{ p \in \mathbb{C}^{n\times k}\;\middle|\;X^{\mathrm{H}}p=-p^{\mathrm{H}}X \right\}.$